added thymeleaf- styling now

+++ src/main/java/com/example/task/management/system/controller/TaskController.java
@@ -3,36 +3,56 @@ package com.example.task.management.system.controller;
 import com.example.task.management.system.model.Task;
 import com.example.task.management.system.service.TaskService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.Optional;
 
-@RestController
-@RequestMapping("/api/tasks")
+@Controller
+@RequestMapping("/tasks")
 public class TaskController {
     @Autowired
     private TaskService taskService;
 
     @GetMapping
-    public List<Task> getAllTasks() {
-        return taskService.getAllTasks();
+    public String getAllTasks(Model model) {
+        List<Task> tasks = taskService.getAllTasks();
+        model.addAttribute("tasks", tasks);
+        return "task-list";
     }
-    @GetMapping("/{id}")
-    public Optional<Task> getTaskById(@PathVariable Long id) {
-        return taskService.getTaskById(id);
+    @GetMapping("/new")
+    public String createTask(Model model) {
+        Task task = new Task();
+        model.addAttribute("task", task);
+        return "task-form";
     }
-    @PostMapping
-    public Task createTask(@RequestBody Task task) {
-        return taskService.createTask(task);
+    @GetMapping("/edit/{id}")
+    public String editTask(@PathVariable Long id, Model model) {
+        Optional<Task> task = taskService.getTaskById(id);
+        if (task.isPresent()) {
+            model.addAttribute("task", task.get());
+            return "task-form";
+        }else{
+            return "redirect:/tasks";
+        }
     }
-    //this might need some help
-    @PutMapping("/{id}")
-    public Task updateTask(@PathVariable Long id, @RequestBody Task task) {
-        return taskService.updateTask(id, task);
+    @PostMapping("/save")
+    public String saveTask(@ModelAttribute("task") Task task) {
+        if (task.getId() != 0){
+            taskService.updateTask(task.getId(), task);
+        }else{
+            taskService.createTask(task);
+        }
+        return "redirect:/tasks";
     }
-    @DeleteMapping("/{id}")
-    public void deleteTask(@PathVariable Long id) {
-        taskService.deleteTaskById(id);
+    @GetMapping("/delete/{id}")
+    public String deleteTask(@PathVariable Long id) {
+        Optional<Task> task = taskService.getTaskById(id);
+        if (task.isPresent()) {
+            taskService.deleteTaskById(id);
+        }
+        return "redirect:/tasks";
     }
 }
